+++ tests/pure/esnext.iterator.constructor.js
@@ -13,6 +13,10 @@ QUnit.test('Iterator', assert => {
   assert.throws(() => Iterator(), 'throws w/o `new`');
 });
 
+QUnit.test('Iterator#constructor', assert => {
+  assert.strictEqual(Iterator.prototype.constructor, Iterator, 'Iterator#constructor is Iterator');
+});
+
 QUnit.test('Iterator#@@toStringTag', assert => {
   assert.strictEqual(Iterator.prototype[Symbol.toStringTag], 'Iterator', 'Iterator::@@toStringTag is `Iterator`');
 });