+++ tests/tests/esnext.async-iterator.constructor.js
@@ -19,6 +19,10 @@ QUnit.test('AsyncIterator', assert => {
   assert.throws(() => AsyncIterator(), 'throws w/o `new`');
 });
 
+QUnit.test('AsyncIterator#constructor', assert => {
+  assert.strictEqual(AsyncIterator.prototype.constructor, AsyncIterator, 'AsyncIterator#constructor is AsyncIterator');
+});
+
 QUnit.test('AsyncIterator#@@toStringTag', assert => {
   assert.strictEqual(AsyncIterator.prototype[Symbol.toStringTag], 'AsyncIterator', 'AsyncIterator::@@toStringTag is `AsyncIterator`');
   assert.strictEqual(String(AsyncIterator.from([1, 2, 3])), '[object AsyncIterator]', 'correct stringification');